Fix rapids CI
